+++ src/smartest/LiteralNode.java
@@ -19,7 +19,7 @@ public class LiteralNode extends ASTNode {
 			int result=value.intValue();
 			this.setIvalue(result);
 		}
-		else if(type.equalsIgnoreCase("double"))
+		else if(type.equalsIgnoreCase("float"))
 		{
 			Double value=(Double)object;
 			double result= value.doubleValue();
@@ -58,7 +58,7 @@ public class LiteralNode extends ASTNode {
 		{
 			output.append(this.getIvalue());
 		}
-		else if(type.equalsIgnoreCase("double"))
+		else if(type.equalsIgnoreCase("float"))
 		{
 			output.append(this.getDvalue());
 		}
